+++ src/main/java/ro/cristiansterie/vote/entity/VoteDAO.java
@@ -13,29 +13,29 @@ public class VoteDAO {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @Column(name = "candidate")
-    private Integer candidateID;
+    private Long candidateID;
     private PartyTypeEnum party;
 
     private Long timestamp;
 
-    private Integer electionId;
+    private Long electionId;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getCandidateID() {
+    public Long getCandidateID() {
         return candidateID;
     }
 
-    public void setCandidateID(Integer candidateID) {
+    public void setCandidateID(Long candidateID) {
         this.candidateID = candidateID;
     }
 
@@ -55,11 +55,11 @@ public class VoteDAO {
         this.timestamp = timestamp;
     }
 
-    public Integer getElectionId() {
+    public Long getElectionId() {
         return electionId;
     }
 
-    public void setElectionId(Integer electionId) {
+    public void setElectionId(Long electionId) {
         this.electionId = electionId;
     }
 }
