+++ src/main/java/ro/cristiansterie/vote/repository/CandidateRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ro.cristiansterie.vote.entity.CandidateDAO;
 
-public interface CandidateRepository extends JpaRepository<CandidateDAO, Integer> {
-    List<CandidateDAO> findAllByElectionId(Integer electionId);
+public interface CandidateRepository extends JpaRepository<CandidateDAO, Long> {
+    List<CandidateDAO> findAllByElectionId(Long electionId);
 
-    int countByElectionId(Integer electionId);
+    int countByElectionId(Long electionId);
 
-    void deleteByElectionId(Integer electionId);
+    void deleteByElectionId(Long electionId);
 }
