+++ src/main/java/ro/cristiansterie/vote/repository/ElectionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ro.cristiansterie.vote.entity.ElectionDAO;
 
-public interface ElectionRepository extends JpaRepository<ElectionDAO, Integer> {
+public interface ElectionRepository extends JpaRepository<ElectionDAO, Long> {
 
     public Optional<ElectionDAO> findFirstByEnabledTrueOrderByStartDateDesc();
 }
