+++ src/main/java/ro/cristiansterie/vote/repository/EventRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ro.cristiansterie.vote.entity.EventDAO;
 
-public interface EventRepository extends JpaRepository<EventDAO, Integer> {
+public interface EventRepository extends JpaRepository<EventDAO, Long> {
 }
